[WIP] Adding baseline of plugin
Currently, this plugin is a rough draft of the end product I am aiming for, with additional functions to be added.

For the time being, it adds support for acr/dockerhub/ecr/gcr but has not been tested. This will need to be tested on each provider.

On next commits I will add Buildkite as a provider, as well as potentially package cloud, as it's more or less a copy and paste job with the way I've designed this to be reusable for the most-part and providing our solutions here is logical for an in-house plugin.

I will move on to testing functionality to ensure the base foundations are working as expected and make changes from there to get it more in line with my vision.

[WIP]